src/main/CinematicFront/src/app/Componentes/home/home.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { HomeComponent } from './home.component';

describe('HomeComponent', () => {
  let component: HomeComponent;
  let fixture: ComponentFixture<HomeComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [HomeComponent]
    })
      .compileComponents();

    fixture = TestBed.createComponent(HomeComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should load popular movies with backdrop paths', () => {
    expect(component.popularMovies).toBeDefined();
    expect(component.popularMovies.length).toBeGreaterThan(0);
    expect(component.popularMovies[0].backdrop_path).toContain('https://');
  });
});
